+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,26 @@
 ## NEXT
 
+- Render GraphQL error debug details and error code in the execution panel nicely. (#9769)
+
+## 2.0.1
+
+- Update internal `firebase-tools` dependency to 15.3.1
+- Fix the data connect emulator discovery bugs.
+
+## 2.0.0
+
+- Update internal `firebase-tools` dependency to 15.0.0
+- [Fixed] common code lens rendering issue.
+  - Don't show "Generate query" in schema file.
+  - Don't show "Generate query" for comments within an operation.
+  - Show "Add Data" for @table type only.
+  - Show "View Data" for @table and @view type.
+- [Fixed] Integration test for recent variables panel changes
+
+## 1.10.2
+
+- Update internal `firebase-tools` dependency to 14.27.0
+
 ## 1.10.1
 
 - Update internal `firebase-tools` dependency to 14.24.2